Travis CI: add sanity check for libelf dependency

+++ travis-ci/managers/debian.sh
@@ -13,6 +13,10 @@ function info() {
     echo -e "\033[33;1m$1\033[0m"
 }
 
+function error() {
+    echo -e "\033[31;1m$1\033[0m"
+}
+
 function docker_exec() {
     docker exec $ENV_VARS -it $CONT_NAME "$@"
 }
@@ -48,6 +52,12 @@ for phase in "${PHASES[@]}"; do
             docker_exec mkdir build
             docker_exec ${CC:-cc} --version
             docker_exec make CFLAGS="${CFLAGS}" -C ./src -B OBJDIR=../build
+            info "ldd build/libbpf.so:"
+            docker_exec ldd build/libbpf.so
+            if ! docker_exec ldd build/libbpf.so | grep -q libelf; then
+                error "No reference to libelf.so in libbpf.so!"
+                exit 1
+            fi
             docker_exec rm -rf build
             ;;
         RUN_ASAN|RUN_CLANG_ASAN|RUN_GCC8_ASAN)
@@ -62,6 +72,12 @@ for phase in "${PHASES[@]}"; do
             docker_exec mkdir build
             docker_exec ${CC:-cc} --version
             docker_exec make CFLAGS="${CFLAGS}" -C ./src -B OBJDIR=../build
+            info "ldd build/libbpf.so:"
+            docker_exec ldd build/libbpf.so
+            if ! docker_exec ldd build/libbpf.so | grep -q libelf; then
+                error "No reference to libelf.so in libbpf.so!"
+                exit 1
+            fi
             docker_exec rm -rf build
             ;;
         CLEANUP)
